Deep Learning 2/requirements.txt
numpy==1.23.5
pandas==1.5.3
matplotlib==3.7.1
scikit-learn==1.2.2
seaborn==0.13.2
tensorflow==2.12.0
keras-tuner==1.4.7
shap
transformers==4.52.4
datasets==3.6.0
torch==2.7.0
torchvision==0.22.0
torchaudio==2.7.0
fastapi==0.115.12
python_multipart==0.0.20
uvicorn==0.34.3
streamlit==1.45.1
